perf(app): derive tokio runtime width from the host; widen the mark price to f64; justify the spot-1m bare key (#1801)
* docs(app): state the single-segment proof the bare-SecurityId key requires

PLAN-EXEMPT: doc comment only, one file, no logic change

`SidServedTracker::per_sid` is a `HashMap<SecurityId, _>` with no
`ExchangeSegment`. `security-id-uniqueness.md` rule 2 permits that ONLY where
the caller can prove every entry belongs to one segment, and requires the
proof to be written at the declaration rather than left for a reader to
reconstruct. The proof was never written.

It does hold, and I traced it rather than assuming it: the only path into
`record_minute` is `record_sid_served_verdicts`, whose verdicts are pushed
inside a `for (slot, (security_id, symbol)) in SPOT_1M_REST_INDICES` loop —
a 4-element const of NIFTY 13, BANKNIFTY 25, SENSEX 51 and INDIA VIX, every
one IDX_I. So the segment half of the composite is a constant and a pair
would add a field that cannot vary.

Two things the comment now carries that the code alone did not:

The same fact BOUNDS the map. The struct doc said "session-scoped", which
reads as "grows during a session and resets" — it cannot exceed 4 keys, ever.
That is a stronger statement and worth having, because CLAUDE.md's O(1) table
records this file's siblings being found unbounded five separate times.

And the condition under which the comment stops being true is stated, so a
future change that feeds this tracker another segment has something to fail
against instead of silently sharing a slot.

Found by a workspace sweep of long-lived per-entity state against the O(1)
table. Same omission class the table already records for
`tick_gap_tracker::states` ("no I-P1-11 justification comment, which the
security-id rule requires"), and the exact inverse of the two maps in
`dhan_universe.rs`, which carry theirs.

Verified: fmt clean, clippy clean, app suite 83 groups green zero failures.
Test count 10345 -> 10352 (the rise is #1798's, not this change's — this
adds no test).


* perf(app): derive tokio runtime width from the host; widen the mark price to f64

Two claims that outlived the facts underneath them, with nothing mechanical
checking either. Both were found by reading the code against its own comments.

ITEM 11a — the runtime width was a constant chosen for one machine.

DEFAULT_TOKIO_WORKER_THREADS = 2 was picked for the 4-vCPU prod box and applied
everywhere: a 1-core dev container got 2 workers (oversubscribed), a larger host
got 2 (under-used), and nothing said which. Memory is already derived this way
in two places -- market_ram_store_boot takes the smaller of /proc/meminfo and
the cgroup limit, and the frame ring reads /proc/meminfo for its 2% -- so CPU
was the one resource still answered by a hardcode.

host_cpu_allowance() asks the cgroup quota FIRST, because that is what the
systemd unit's AllowedCPUs actually confines this process to;
available_parallelism() reports the MACHINE, which on a confined process is the
wrong, larger number. The result is clamped into [1, 4]: 0 would panic
Builder::worker_threads at boot, and a bigger box must not silently widen the
runtime past what the CPU isolation intends. The env override still wins, so
the operator's rollback path is unchanged -- config plus a restart, no rebuild.

The derivation is emitted with its SOURCE (cgroup / parallelism / fallback),
not just its answer: a derived number nobody can trace back to its input is the
thing this change set out to remove. It is stashed in a OnceLock and logged
from async_main because main runs before the subscriber exists.

ITEM 11b — MarkUpdate.price was f32 on a premise that died three months ago.

The field doc said "f32 as delivered", true while a live WS bridge delivered
f32 ticks, and false since that source was removed. Every remaining caller is a
REST leg delivering f64 (candle.close; leg.last_price), and the far end,
RiskEngine::update_market_price_in_segment, takes f64 and feeds the daily-loss
auto-halt. So the pipe was f64 at both ends with a narrowing in the middle that
bought nothing.

It cost precision above 131,072, where f32's step is 0.0156 -- coarser than a
paisa. No price flowing today reaches that, so this removes a latent class
rather than fixing a live defect, and the record says so.

One test changed and it is worth naming. test_mark_update_is_copy_and_small
pinned size_of <= 16 and FAILED: u64 + u8 + f64 pads to 24. That ratchet did
its job. The bound moved only after the cost was computed -- the mark channel
is bounded at mark_channel_capacity (default 8,192, max 65,536), so 8 bytes per
slot is +64 KiB at the default and +512 KiB at the ceiling, on a 32 GiB host.
The contract the test protects, a Copy payload passed by value with no heap
behind it, is intact.

Measured, not claimed:
  cargo fmt --check                                     clean
  cargo clippy --workspace --no-deps -- -D warnings     clean (the CI gate)
  cargo test -p tickvault-app                           83 suites, 0 failures
  dhat_mark_forward_hot_arms_zero_allocation            pass
  order_gate/mark_forward_disarmed                      2.49 ns  (budget 100)
  order_gate/mark_forward_armed_full                    9.24 ns  (budget 100)
  order_gate/mark_forward_armed_accept                  69.9 ns  (budget 100)
  tv-guarantees                                         22/13/2/0, exit 0

What this does NOT deliver: 11a makes the width follow the host; it does not
measure whether 2, 3 or 4 workers is right for this workload. The first live
session at scale is still that measurement.


---------

+++ crates/app/src/dhan_cadence_executor.rs
@@ -623,11 +623,7 @@ impl CadenceExecutor for DhanCadenceExecutor {
                 // minutes late is exactly the staleness the decision-freshness
                 // gate forbids.
                 if let Some(forwarder) = self.mark_forwarder.as_ref() {
-                    forwarder.mark_forward(
-                        security_id,
-                        EXCHANGE_SEGMENT_IDX_I,
-                        candle.close as f32,
-                    );
+                    forwarder.mark_forward(security_id, EXCHANGE_SEGMENT_IDX_I, candle.close);
                 }
             }
             metrics::counter!("tv_spot1m_fetch_total", "outcome" => "ok").increment(1);
@@ -991,7 +987,7 @@ impl CadenceExecutor for DhanCadenceExecutor {
                                 refused_zero_id += 1;
                                 continue;
                             };
-                            forwarder.mark_forward(contract_id, seg_code, leg.last_price as f32);
+                            forwarder.mark_forward(contract_id, seg_code, leg.last_price);
                         }
                         if refused_zero_id > 0 {
                             metrics::counter!(
